feat(refactor): move logoug sagas to session
