+++ packages/blockchain-wallet-v4-frontend/src/data/auth/sagaRegister.ts
@@ -7,10 +7,7 @@ export default ({ api, coreSagas, networks }) => {
   const authSagas = sagas({ api, coreSagas, networks })
 
   return function* authSaga() {
-    yield takeLatest(actions.deauthorizeBrowser.type, authSagas.deauthorizeBrowser)
     yield takeLatest(actions.login.type, authSagas.login)
-    yield takeLatest(actions.logout.type, authSagas.logout)
-    yield takeLatest(actions.logoutClearReduxStore.type, authSagas.logoutClearReduxStore)
     yield takeLatest(actions.mobileLogin.type, authSagas.mobileLogin)
     yield takeLatest(actions.register.type, authSagas.register)
     yield takeLatest(actions.resetAccount.type, authSagas.resetAccount)
