+++ packages/blockchain-wallet-v4-frontend/src/data/session/slice.ts
@@ -9,6 +9,9 @@ const sessionSlice = createSlice({
   initialState,
   name: 'session',
   reducers: {
+    deauthorizeBrowser: () => {},
+    logout: () => {},
+    logoutClearReduxStore: () => {},
     removeSession: (state, action: PayloadAction<string>) => {
       return dissoc(action.payload, state)
     },
@@ -18,7 +21,8 @@ const sessionSlice = createSlice({
   }
 })
 
-export const { removeSession, saveSession } = sessionSlice.actions
+export const { deauthorizeBrowser, logout, logoutClearReduxStore, removeSession, saveSession } =
+  sessionSlice.actions
 
 const { actions } = sessionSlice
 const sessionReducer = sessionSlice.reducer
