+++ packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/MenuTop/Settings/index.tsx
@@ -9,8 +9,8 @@ import Settings from './template'
 const SettingsContainer = (props: Props) => <Settings {...props} />
 
 const mapDispatchToProps = (dispatch) => ({
-  authActions: bindActionCreators(actions.auth, dispatch),
   modalActions: bindActionCreators(actions.modals, dispatch),
+  sessionActions: bindActionCreators(actions.session, dispatch),
   settingsActions: bindActionCreators(actions.modules.settings, dispatch)
 })
