+++ packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/MenuTop/Settings/template.tsx
@@ -99,7 +99,7 @@ const Settings = (props: Props) => {
             <DropdownSeparator />
             <DropdownMenuItem
               onClick={() => {
-                props.authActions.logout()
+                props.sessionActions.logout()
               }}
               data-e2e='logoutLink'
             >
